requirements.txt
Flask>=1.0,<=1.1.2
pyodbc==4.0.32
azure-storage-blob==12.9.0
requests==2.26.0
cognitive-face==1.5.0
Pillow==8.3.2
azure-ai-textanalytics==5.1.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-vision-face==0.5.0
Flask-Bootstrap==3.3.7.1
Flask-Moment==1.0.2
Flask-WTF==0.15.1
Flask-Login==0.5.0
sendgrid==6.8.2
SQLAlchemy == 1.4.25